+++ piwigo/Piwigo Data Wrappers/ImageService.m
@@ -524,7 +524,7 @@ NSString * const kGetImageOrderDescending = @"desc";
                                           NSInteger count = [albumData addImages:albumImages];
 #if defined(DEBUG_LOAD)
                                           NSInteger imagesPerPage = [ImagesCollection numberOfImagesToDownloadPerPage];
-                                          NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %lu)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, albumData.imageList.count % imagesPerPage);
+                                          NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %ld)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, (long)(albumData.imageList.count % imagesPerPage));
 #endif
                                           if (completion) {
                                               completion(task, count);
@@ -556,7 +556,7 @@ NSString * const kGetImageOrderDescending = @"desc";
                                           NSInteger count = [albumData addImages:albumImages];
 #if defined(DEBUG_LOAD)
                                           NSInteger imagesPerPage = [ImagesCollection numberOfImagesToDownloadPerPage];
-                                          NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %lu)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, albumData.imageList.count % imagesPerPage);
+                                          NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %ld)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, (long)(albumData.imageList.count % imagesPerPage));
 #endif
                                           if (completion) {
                                               completion(task, count);
@@ -585,7 +585,7 @@ NSString * const kGetImageOrderDescending = @"desc";
                                             NSInteger count = [albumData addImages:albumImages];
 #if defined(DEBUG_LOAD)
                                             NSInteger imagesPerPage = [ImagesCollection numberOfImagesToDownloadPerPage];
-                                            NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %lu)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, albumData.imageList.count % imagesPerPage);
+                                            NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %ld)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, (long)(albumData.imageList.count % imagesPerPage));
 #endif
                                             if (completion) {
                                                 completion(task, count);
@@ -617,7 +617,7 @@ NSString * const kGetImageOrderDescending = @"desc";
                                               NSInteger count = [albumData addImages:albumImages];
 #if defined(DEBUG_LOAD)
                                               NSInteger imagesPerPage = [ImagesCollection numberOfImagesToDownloadPerPage];
-                                              NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %lu)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, albumData.imageList.count % imagesPerPage);
+                                              NSLog(@"loadImageChunkFor…: +%ld img in cat: %ld (%ld/%ld : %ld)", (long)count, (long)categoryId, (long)albumData.imageList.count, (long)albumData.numberOfImages, (long)(albumData.imageList.count % imagesPerPage));
 #endif
                                               if (completion) {
                                                   completion(task, count);
